requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.14.0
google-generativeai>=0.3.0

# Data & Technical Analysis
pandas>=2.0.0
pandas-ta==0.3.14b; python_version >= "3.12"
ta>=0.10.0
numpy>=1.24.0
yfinance>=0.2.0

# Web Server (Render)
flask>=3.0.0
gunicorn>=21.0.0

# Firebase
requests>=2.31.0

# Utilities
beautifulsoup4>=4.12.0
schedule>=1.2.0
Pillow>=10.0.0
